add js file minifying step to ci, move directory
