+++ .github/workflows/pi-embed-actions-ci.yml
@@ -1,0 +1,44 @@
+name: PiEmbedActions CI Pipeline
+on:
+  push:
+    branches:
+      - '**'   # This will trigger on pushes to any branch
+  pull_request:
+    branches:
+      - '**'   # This will trigger on pull requests to any branch
+
+jobs:
+  common:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v2
+      - name: Set up Node.js
+        uses: actions/setup-node@v2
+        with:
+          node-version: '18'
+      - name: Install dependencies
+        run: npm install
+
+  test:
+    runs-on: ubuntu-latest
+    needs: common
+    steps:
+      - name: Run js tests
+        run: npm test
+
+  build:
+    runs-on: ubuntu-latest
+    needs:
+      - common
+      - test
+   # if: github.ref == 'refs/heads/main'  
+
+    steps:
+      - name: Minify and Bundle the library
+        run: npm run build
+      - name: Upload minified version as artifact
+        uses: actions/upload-artifact@v2
+        with:
+          name: pi-embed-actions-v1.0.0.min
+          path: dist/pi-embed-actions-v1.0.0.min.js
