--- .github/workflows/test.yml
@@ -1,22 +1,0 @@
-name: Run JS Tests
-on:
-  push:
-    branches:
-      - '*' 
-  pull_request:
-    branches:
-      - '*'
-jobs:
-  test:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v3
-      - name: Set up Node.js
-        uses: actions/setup-node@v3
-        with:
-          node-version: 18
-      - name: Install dependencies
-        run: npm install
-      - name: Run js tests
-        run: npm test
