+++ examples/example.js
@@ -1,4 +1,4 @@
-import PiEmbedActions from './main.js';
+import PiEmbedActions from '../src/main.js';
 
 const dashboard2 = PiEmbedActions.createDashboard('dashboard2', 'https://pi-dev.uk:8224/pi?lang=en_GB&editorDisplayMode=CONTENT');
 dashboard2.reportEditor.open(123);
