+++ src/internal/report-editor.js
