+++ src/internal/utils/editor-display-mode.js
