+++ src/internal/utils/embed-utils.js
